Store auth restructure
